Lecture profile dans un fichier JSON

+++ app/src/main/java/com/michelAdrien/AMTT0121/Model/UserProfile.kt
@@ -1,5 +1,5 @@
 package com.michelAdrien.AMTT0121.Model
 
-class UserProfile(val firstName : String, val lastName:String, val address : Adress, val BirthDate:Long) {
+class UserProfile(val firstName : String, val lastName:String, val address : Adress, val BirthDate:Double) {
 
 }
